+++ src/controllers/MainCategoryController.php
@@ -173,7 +173,7 @@ class MainCategoryController extends Controller {
 		try {
 			$main_category = MainCategory::withTrashed()->where('id', $request->id)->first();
 			if (!is_null($main_category->icon_id)) {
-				Attachment::where('attachment_of_id', 21)->where('entity_id', $request->id)->forceDelete();
+				Attachment::where('company_id', Auth::user()->company_id)->where('attachment_of_id', 21)->where('entity_id', $request->id)->forceDelete();
 			}
 			MainCategory::withTrashed()->where('id', $request->id)->forceDelete();
